moved relevant files to server folder
